Rusty Skylines Alpha v1.7.2a April 20th 22:22
[Major]
Added Lots, which are the actual lots for houses.
These lots have a width and length.
You go into the zoning mode, then you will see lots of points on the sides of the roads.
These points are where you can place the lot against.
After you placed a lot, you can change its zone_type by simply clicking on the Lot with your desired ZoneType set on the cursor.
It prevents you from placing if the lot collides with other lots or roads!

[Minor]
Made astronomy more accurate, all use gmst now at 2026 01 01 00:00  Total game time is seconds since this date. Should match up with sun and moon irl and phases too.

Zooming is now resistant to spam clicking and is always smooth because of it! L2 C2 Continuity?

All meshed roads now give me RoadEdges so I can snap to them and test against them like in Lots!

Made RoadType be stored in a RoadTypes storage and be accessed by an ID hash, so that the whole RoadType of 30 Bytes won't have to be cloned and stored in every single road element!
I know, 30 bytes is tiny, but it's for the future and cleaner in code!

The moon direction is now also shown on the gizmo.

The gizmo now only shows when the debug menu is open. Like in MINECRAFT!!

***
I neglected the Area Zoning too from last update!

Important: I gave up on Saving! Saving will only work for as many versions as it can, but it will break and I will care, but it's just way too hard to implement and not worth it!
I will make saving better once Everything in the game is way more stable, like in Beta or even Pre release, it depends and we'll see!

Not too much coding, lots of procrastination, at least I got something!

+++ src/renderer/uniform_updates.rs
@@ -95,15 +95,15 @@ impl<'a> UniformUpdater<'a> {
         );
     }
     pub fn update_sky_uniforms(&self, astronomy: &Astronomy) {
-        let moon_phase: f32 = astronomy.moon_phase;
+        let moon_phase: f32 = astronomy.moon_phase as f32;
 
         let sky_uniform = SkyUniform {
             star_rotation: astronomy.star_rotation.to_cols_array_2d(),
             exposure: 1.0,
             moon_phase,
-            sun_size: astronomy.sun_size_ndc,
+            sun_size: astronomy.sun_size_ndc as f32,
             sun_intensity: 1.0,
-            moon_size: astronomy.moon_size_ndc,
+            moon_size: astronomy.moon_size_ndc as f32,
             moon_intensity: 1.0,
             _pad1: 1.0,
             _pad2: 0.0,
